Integrate using RubyAmazonBedrock gem
This is an example of how to integrate the RubyAmazonBedrock gem
into a Ruby application.

It's a gem owned and maintained by an Amazon developer. It's pretty
awesome and super easy to use/configure options etc.

I'll add a second implementation without using this
gem as it looks like it's maintained by a single developer
so you might not want to use it in prod.